storage v1 interactive mode
