+++ tkcore/term/storage_v2/storage_v2.h
@@ -5,7 +5,7 @@
 #ifndef THEKEY_STORAGE_V2_H
 #define THEKEY_STORAGE_V2_H
 
-#include "def_header.h"
+#include "../def_header.h"
 
 namespace thekey_term_v2 {
